refactor: Update user name to userName and add firstName and lastName fields in UserInfos component
This commit updates the UserInfos component to use the "userName" field instead of the "name" field for the username input. It also adds new inputs for the "firstName" and "lastName" fields. This change ensures consistency in the usage of user information and improves the user experience by allowing users to provide their first and last names.